Move AllyGen to Allies namespace
