+++ PyRoute/Calculation/CommCalculation.py
@@ -5,7 +5,7 @@ Created on Aug 09, 2023
 """
 import networkx as nx
 
-from PyRoute.AllyGen import AllyGen
+from PyRoute.Allies.AllyGen import AllyGen
 from PyRoute.Calculation.RouteCalculation import RouteCalculation
 try:
     from PyRoute.Pathfinding.ApproximateShortestPathForestUnified import ApproximateShortestPathForestUnified
