+++ PyRoute/Calculation/OwnedWorldCalculation.py
@@ -3,7 +3,7 @@ Created on Aug 09, 2023
 
 @author: CyberiaResurrection
 """
-from PyRoute.AllyGen import AllyGen
+from PyRoute.Allies.AllyGen import AllyGen
 from PyRoute.Calculation.RouteCalculation import RouteCalculation
