+++ PyRoute/DeltaDebug/DeltaDictionary.py
@@ -9,7 +9,7 @@ import logging
 import os
 
 from PyRoute.DeltaDebug.DeltaLogicError import DeltaLogicError
-from PyRoute.AllyGen import AllyGen
+from PyRoute.Allies.AllyGen import AllyGen
 from PyRoute.AreaItems.Allegiance import Allegiance
 from PyRoute.AreaItems.Sector import Sector
 from PyRoute.Star import Star
