+++ PyRoute/DeltaDebug/DeltaGalaxy.py
@@ -4,13 +4,13 @@ Created on May 21, 2023
 @author: CyberiaResurrection
 """
 
+from PyRoute.Allies.AllyGen import AllyGen
 from PyRoute.AreaItems.Allegiance import Allegiance
 from PyRoute.AreaItems.Galaxy import Galaxy
 from PyRoute.AreaItems.Subsector import Subsector
 from PyRoute.AreaItems.Sector import Sector
 from PyRoute.DeltaDebug.DeltaDictionary import SectorDictionary, SubsectorDictionary
 from PyRoute.Star import Star
-from PyRoute.AllyGen import AllyGen
 
 
 class DeltaGalaxy(Galaxy):
